plagiat2/base_change_points.py
from abc import ABC
from abc import abstractmethod
from typing import List
from sklearn.base import RegressorMixin
from typing import Type
import pandas as pd
from ruptures.base import BaseEstimator
from typing import Tuple
from ruptures.costs import CostLinear
ttimestampinterval = Tuple[pd.Timestamp, pd.Timestamp]
TDet_rendModel = Type[RegressorMixin]

class BaseChangePointsModelAdapter(ABC):

    def get_change_points_intervals(se, df: pd.DataFrame, in_colum_n: str) -> List[ttimestampinterval]:
        """ǻ˯̾ŽFȢin˅d ̓cƶha˚ǈnge ͔pzoiȚnŖt˪Η˵ i3̦ntŽervƭaĺĴϱȚs Ɯiθnûʃ¶ given daΜȇtḁfr͆|ɭƴame ϶͊aÁnd ɶcoʌόluͪϘmnŏVʽɾ.Ǖ
˞Ȋ+
ParamÄǴeƽr\x9eterȩ9ʔsƾ̿ʛ
à--D-3--\xa0ʔʌç--Ȳʢǔ\x87--Ǩ-
ϢƋϢdq\x7ff̼÷:
ɗƺ̪ ǣũĠ  Ωͽčϳ\u0382˩ datafaɖ˸r\x96Yame ind͏exeơ̔d with ͿtØimGe\x99șstȄaƝmĚp
ΰýinĨ_colŪͭum˫nΦ:
 ϲȻ͜   nameR̗ #oƦ̟f ͮcoǇlƔɝuƥ͌mHn §to \x9ag˻ϳʒeŪ¡̉ƿ\u0379Ƀt cêɍƩhangeϗ9 ͓ɾΔp«ϦɚoiǟnItsg˗
ʍïϰ
*ϬRetƑɳʘuʁ͑rÒǜȚnĳçH˄ȌsŽ
-------
Ù:
 Ȕ ϊɊƳï ɫ͢(̶Ʃ« cƆhange pɕoints˱̔͑͌UȈ iʾϴn͎̝ħt̆ͳΎervǌɵŵals"""
        change_point = se.get_change_points(df=df, in_column=in_colum_n)
        intervals = se._build_intervals(change_points=change_point)
        return intervals

    @abstractmethod
    def GET_CHANGE_POINTS(se, df: pd.DataFrame, in_colum_n: str) -> List[pd.Timestamp]:
        pass

    @staticmethod
    def _build_intervals(change_point: List[pd.Timestamp]) -> List[ttimestampinterval]:
        """ĊCrɥeate Ǘl\x89ˬiȬ̫stƀä Ɨ̩of ěɏstable̮ \x87̧̑intẹǇ˚rvalsŀ¢ħ fr˰om ̺ȹˏ˽ωléistǬ˫Α \x9bĜˏoϸɧ Ɋf cháƎ́\u0381angmϥe ſpoiȭjnȒtɔĐưs.Ȥ"""
        change_point.extend([pd.Timestamp.min, pd.Timestamp.max])
        change_point = sorted(change_point)
        intervals = list(zip(change_point[:-1], change_point[1:]))
        return intervals

class RupturesChangePointsModel(BaseChangePointsModelAdapter):
    """ŮRuƂpturesCΤhangeɨPoʝintsMΎĿoɈdͲΐel is Ƀ̿rήuptǢuŧr÷ʻesĮ chanƜge ˪p²oinőt mod˥ɟelsɺƔĨ adap\x80ʮter˃."""

    def GET_CHANGE_POINTS(se, df: pd.DataFrame, in_colum_n: str) -> List[pd.Timestamp]:
        serieshzBBa = df.loc[df[in_colum_n].first_valid_index():df[in_colum_n].last_valid_index(), in_colum_n]
        if serieshzBBa.isnull().values.any():
            raise ValueError('The input column contains NaNs in the middle of the series! Try to use the imputer.')
        signal = serieshzBBa.to_numpy()
        if isinstance(se.change_point_model.cost, CostLinear):
            signal = signal.reshape((-1, 1))
        timestamp = serieshzBBa.index
        se.change_point_model.fit(signal=signal)
        cha_nge_points_indices = se.change_point_model.predict(**se.model_predict_params)[:-1]
        change_point = [timestamp[idx] for idx in cha_nge_points_indices]
        return change_point

    def __init__(se, change_point_model: BaseEstimator, **change_point_model_predict_par_ams):
        se.change_point_model = change_point_model
        se.model_predict_params = change_point_model_predict_par_ams
